nativedsl: pool macro params and for-bindings as ChildRange
Convert MacroConfig.params and ForConfig.bindings from Vec<Param> to a
ChildRange into a new shared AstPools.params pool, removing one heap
allocation per macro and per for-loop. push_params returns ChildRange
directly (not Option): both counts are already bounded to u8::MAX at parse
time (their indices are u8), so the length always fits ChildRange::len.

Pure representation change; no behavior change.

+++ crates/generate/src/nativedsl/lower/evaluator.rs
@@ -1159,7 +1159,7 @@ impl<'a, 'ast> Evaluator<'a, 'ast> {
     {
         let for_config = self.shared.pools.get_for(for_id);
         let iterable = for_config.iterable;
-        let n_bindings = for_config.bindings.len();
+        let n_bindings = usize::from(for_config.bindings.len);
         let iter_vid = self.eval_expr(iterable)?;
         let iter_range = self.list_range(iter_vid);
         let n_items = usize::from(iter_range.len);
